psg: Fix frequency divisor (it sounds good now!)
The PSG divides the master clock by 16, and we previously clocked directly at
the master frequency.

Went back using doubles for PSG_CLOCKS_PER_SAMPLE.  The current value is
5.07305130385, so rounded down to 5 the difference is not huge, but the pitches
sound better with the extra precision.  (Besides, it /is/ the correct thing to
do).

Now Sonic 1 Stage 1 sounds okay!  The PSG is used for accompanying notes though,
not the main melody: that's the YM2612.

+++ megado/psg.h
@@ -27,8 +27,8 @@ typedef struct NoiseChannel {
 } NoiseChannel;
 
 typedef struct PSG {
-    uint16_t remaining_clocks;
-    uint32_t sample_counter;
+    int16_t remaining_clocks;
+    double sample_counter;
 
     SquareChannel square[3];
     NoiseChannel noise;
